refactor: 테스트 파일 및 경로 정리, setup 파일 변경
