+++ src/domains/ShoppingProducts/apis/cartItem.ts
@@ -1,4 +1,4 @@
-import { apiClient } from "../utils/apiUtils";
+import { apiClient } from "./utils/apiUtils";
 
 export async function getCartItems({ sortBy }: { sortBy: string }) {
   const params = new URLSearchParams({
